+++ jscomp/build_tests/custom_namespace/input.js
@@ -4,7 +4,7 @@ var path = require("path");
 var assert = require("assert");
 var rescript_exe = require("../../../scripts/bin_path").rescript_exe;
 
-child_process.execSync(`${rescript_exe} clean -with-deps && ${rescript_exe} build`, {
+child_process.execSync(`${rescript_exe} clean && ${rescript_exe} build`, {
   cwd: __dirname,
 });