+++ jscomp/build_tests/uncurried-always/input.js
@@ -5,5 +5,5 @@ const fs = require('fs')
 const path = require('path')
 var rescript_exe = require("../../../scripts/bin_path").rescript_exe
 
-cp.execSync(`${rescript_exe} clean -with-deps`, { cwd: __dirname, });
+cp.execSync(`${rescript_exe} clean`, { cwd: __dirname, });
 cp.execSync(`${rescript_exe}`, { cwd: __dirname, });